feat(core): deprecate ui.forElement and ui.forElements
The `ui.forElement`and `ui.forElements` functions are deprecated and will be removed in a future release. Please use `uiFor` function provided as a property to your element ui implementation.

Each element in the UI tree must refer to an *unchanged* element in the state tree. Allowing arbitrary calls to `forElement(s)` allows the users to transform elements in an e.g. parent node before passing them down to child nodes, so they are deprecated.

Imposing this restriction allows us to optimize the re-rendering calls to just the sub-tree where the change had happened. This is an important optimization as a top-level elements often use 3rd party heavy components (e.g. react-navigation) which aren't optimized for frequent re-rendering like ordinary element UI is.