fix(components): Improve viewport awareness higher-order component (#62)
* Optimise layout measurements of viewport aware components

This commit will introduce an optimisation that will greatly improve
the performance of the onViewportChange calculations.

* Fix setting of state on unmounted viewport aware components

This commit will make sure that the state of unmounted components is
never set. This would sometimes happen when the UIManager.measureLayout
callback gets invoked and the component has already unmounted.

* Improve readability of onViewportChange

This commit will improve the readability of the code in the function.