Enhancers (an alternative to Enrichers) (#87)
Implement *enhancers* as an alternative to *enrichers*.

Profiling showed that there is a significant performance impact of using enrichers mostly due to the inefficient Promise implementation present in react native. Enhancers are a solution (probably a temporary one) that should perform better.

Enhancers, like enrichers are registered for a specific element, but they will be only allowed to act on root level elements.

Enhancers have the API
```javascript
// registration
enhance.register(kind, enhancer)

// enhancer signature
enhancer :: async (el, context) => (el => el)
```

In other words, they take the element, an optional context and then return an updating fn that will eventually apply the enhancement to the element in question.

All enhancers are run in parallel on the read document. The order in which the returned update fns are run is determined by the subsystem sequence / registration sequence.

The idea is that the enhancer performs necessary async / fetch operations in the body of the async call (by examining the provided document), once everything is ready it returns the update fn which will apply the changes on the document.

e.g.

```javascript
enhance.register(['document'], async (el, context) => {
  const bookmarks = await getBookmarks(context)

  return document => applyBookmarks(document)
})
```

Closes #84